requirements.txt
fastapi==0.111.0
ecs-logging==2
aws_embedded_metrics
asgi-logger
pydantic_settings
dnspython
pymongo
pymongo[aws,snappy,zstd,encryption]
httpx

beautifulsoup4
isort

langchain
langchain-chroma
langchain-community
langchain-core
langchain_aws
langchain_openai
langchain_azure_ai

langgraph

anthropic
boto3
openai
